# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Falzoni PHP Laravel",
        "description": "Api de serviços com Laravel PHP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/get/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obter usuario por Id",
                "description": "Retorna o objeto do usuario",
                "operationId": "45cd93f3dc0a2e6021234f8776177f22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitação efetuada com sucesso"
                    }
                }
            }
        },
        "/api/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Adicionar usuário",
                "description": "Cria um novo registro de usuário",
                "operationId": "f457f406a3220d6f646f5874ab56b0a7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "user_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "full_name": "Test User",
                                    "user_name": "testuser",
                                    "email": "testuser@test.com.br",
                                    "password": "abc123",
                                    "phone_number": "(11) 92222-2222"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "description": "Autentica o usuário e retorna o bearer token",
                "operationId": "0d86aa2f7690604b9622c1ba099ac4b9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "testuser@test.com.br",
                                    "password": "abc123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/user/update/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Atualizar usuário",
                "description": "Atualiza um registro de usuário",
                "operationId": "bd2929e09ee6e798edcc639128158681",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "user_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "full_name": "Test User",
                                    "user_name": "testuser",
                                    "email": "testuser@test.com.br",
                                    "password": "abc123",
                                    "phone_number": "(11) 92222-2222"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/user/delete/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Remover usuário por Id",
                "description": "Remove o usuário",
                "operationId": "5f2aa3ebbb303f17e74faef8bd435d10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Token de autorização (JWT) a ser inserido",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}